src/requirements.txt
asgiref==3.4.1
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
colorama==0.4.4
cryptography==3.4.8
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-autoslug==1.9.8
django-extensions==3.1.3
django-jazzmin==2.4.8
django-rest-framework==0.1.0
django-taggit==1.5.1
djangorestframework==3.12.4
<<<<<<< HEAD
djangorestframework-jwt==1.11.0
=======
drf-jwt==1.19.0
>>>>>>> 1ed7da5f399e20a2b7473c81dc9d369b2b27de94
idna==3.2
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.3
oauthlib==3.1.1
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2==2.9.1
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
soupsieve==2.2.1
sqlparse==0.4.2
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
